app/requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.6
Django==3.0.8
django-model-utils==4.0.0
django-notifications-hq==1.6.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
future==0.18.2
importlib-metadata==1.7.0
jsonfield==3.1.0
kombu==4.6.11
Pillow==7.2.0
pytz==2020.1
redis==3.5.3
sqlparse==0.3.1
swapper==1.1.2.post1
vine==1.3.0
XlsxWriter==1.2.9
xlwt==1.3.0
zipp==3.1.0
psycopg2-binary==2.8.5
